job_portal/requirements.txt
asgiref==3.3.4
astroid==2.5.6
Django==3.1.6
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-js-asset==1.2.2
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.1.1
pylint==2.8.2
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
